requirements-poc.txt
pandas>=2.0
openpyxl>=3.1
pyyaml>=6.0
streamlit>=1.33
numpy>=1.24
pytest>=8.0
